BSSGP_Types: Fix coding of Reporting Cell Identifier
3GPP TS 48.018 V10.6.0 (2012-07), section 11.3.63.1 describes the reporting
Cell Identifier IE coding as follows: "Reporting Cell Identifier: This field
is encoded as the Cell Identifier defined in sub-clause 11.3.9", but when
looking further at 11.3.63.2 the Cell Identifier IE is described as
"Reporting Cell Identifier: This field is encoded as the value part of the
Cell Identifier IE defined in sub-clause 11.3.9, not including IEI and Length
Indicator." The latter description makes more sense, since technically the
IE and length of the Cell Identifier from 11.3.9 would be redundant. The
same also applies for section 11.3.63.4.

Change-Id: I7365993366a92db3aa7e3982c6b45d97da225acd